+++ src/tests/libxrpl/protocol/BuildInfo.cpp
@@ -1,0 +1,98 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2020 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#include <xrpl/protocol/BuildInfo.h>
+
+#include <doctest/doctest.h>
+
+using namespace ripple;
+
+TEST_SUITE_BEGIN("BuildInfo");
+
+TEST_CASE("EncodeSoftwareVersion")
+{
+    auto encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.3-b7");
+
+    // the first two bytes identify the particular implementation, 0x183B
+    CHECK(
+        (encodedVersion & 0xFFFF'0000'0000'0000LLU) ==
+        0x183B'0000'0000'0000LLU);
+
+    // the next three bytes: major version, minor version, patch version,
+    // 0x010203
+    CHECK(
+        (encodedVersion & 0x0000'FFFF'FF00'0000LLU) ==
+        0x0000'0102'0300'0000LLU);
+
+    // the next two bits:
+    {
+        // 01 if a beta
+        CHECK((encodedVersion & 0x0000'0000'00C0'0000LLU) >> 22 == 0b01);
+        // 10 if an RC
+        encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.4-rc7");
+        CHECK((encodedVersion & 0x0000'0000'00C0'0000LLU) >> 22 == 0b10);
+        // 11 if neither an RC nor a beta
+        encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.5");
+        CHECK((encodedVersion & 0x0000'0000'00C0'0000LLU) >> 22 == 0b11);
+    }
+
+    // the next six bits: rc/beta number (1-63)
+    encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.6-b63");
+    CHECK((encodedVersion & 0x0000'0000'003F'0000LLU) >> 16 == 63);
+
+    // the last two bytes are zeros
+    CHECK((encodedVersion & 0x0000'0000'0000'FFFFLLU) == 0);
+
+    // Test some version strings with wrong formats:
+    // no rc/beta number
+    encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.3-b");
+    CHECK((encodedVersion & 0x0000'0000'00FF'0000LLU) == 0);
+    // rc/beta number out of range
+    encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.3-b64");
+    CHECK((encodedVersion & 0x0000'0000'00FF'0000LLU) == 0);
+
+    // Check that the rc/beta number of a release is 0:
+    encodedVersion = BuildInfo::encodeSoftwareVersion("1.2.6");
+    CHECK((encodedVersion & 0x0000'0000'003F'0000LLU) == 0);
+}
+
+TEST_CASE("IsRippledVersion")
+{
+    auto vFF = 0xFFFF'FFFF'FFFF'FFFFLLU;
+    CHECK(!BuildInfo::isRippledVersion(vFF));
+    auto vRippled = 0x183B'0000'0000'0000LLU;
+    CHECK(BuildInfo::isRippledVersion(vRippled));
+}
+
+TEST_CASE("IsNewerVersion")
+{
+    auto vFF = 0xFFFF'FFFF'FFFF'FFFFLLU;
+    CHECK(!BuildInfo::isNewerVersion(vFF));
+
+    auto v159 = BuildInfo::encodeSoftwareVersion("1.5.9");
+    CHECK(!BuildInfo::isNewerVersion(v159));
+
+    auto vCurrent = BuildInfo::getEncodedVersion();
+    CHECK(!BuildInfo::isNewerVersion(vCurrent));
+
+    auto vMax = BuildInfo::encodeSoftwareVersion("255.255.255");
+    CHECK(BuildInfo::isNewerVersion(vMax));
+}
+
+TEST_SUITE_END();
